Fix: EVP_Cipher() is supposed to return the output length, or -1 on error
EVP_Cipher() is undocumented, but study of previous OpenSSL versions,
it's clear that it is supposed to return the length of the encrypted
or decrypted content, and that 0 is a permissible output length,
leaving -1 for errors.

The corresponding provider dispatched function has different
semantics, where it returns 0 on error and 1 on success, and sets the
output length to one of it's writable parameters.  We therefore need
to "translate" those returned values to EVP_Cipher() semantics.