Make FIPS provider algorithms selectable
For Configure, we add a generic option to specify the desired
algorithms in a provider.
For the FIPS provider, '--select-fips-algorithms=' is the option, and
its value is a command separated list of algorithms, possibly prefixed
with a '+' or a '-' to specify if an algorithm should be added ot
removed.

The Perl nuggets in providers/fips/fipsprov.c.in are changed to handle
this configuration.

Regarding the selections, the can be specified as algorithm names or
algorithm sets, and may be given just plain, or prefixed with a '+' or
a '-'.  If given in plain, no default will apply.  If prefixed with
'+', that algorithm or algorithm set will be added to the current
selection.  If prefixed with '-', the algorithm or algorithm set will
be removed from the current selection.

Finally, and algorithm set is a name that represent a list of
algorithms.  These are specified in the '.desc' files as en entry with
the value being an array rather than a hash entry.