CORE: Add a generic callback structure and upcall
The idea is that libcrypto code will be able to build up the structure
it needs for any callback, pass that to the provider where the provider
will accept it, and the provider passes that back along with an
OSSL_PARAM array with data from the provider to the core_generic_callback
upcall, which knows how to handle the passed structure.

The structure is very simple, and includes a function pointer to a
libcrypto callback, as well as a pointer to some libcrypto data.  That
libcrypto callback needs to know what to do with that data to call the
application callback with the arguments it requires.

This method permits the provider to call application callbacks without
having any information of what signature they have, and caters to those
who may want to build a provider module into a secure environment where
all the calls simply can't be any random function call, but rather must
be in form of up- or downcalls in the libcrypto<->provider interface.