squash! fixup! Replumbing: pre-populate the EVP namemap with commonly known names
This affected the way legacy NIDs are looked up in evp_cipher_from_dispatch()
and evp_md_from_dispatch().  Instead of trying to find the NID directly,
look up the legacy method structure and grab the NID from there.  The
reason is that NIDs can be aliases for other NIDs, which looked like
a clash even if wasn't really one.