Add OSSL_RET_UNSUPPORTED and use it in obvious places
OSSL_RET_UNSUPPORTED is -2, which is a return value that is used
throughout the OpenSSL source to signify that the operation,
parameters values or combination of parameters or similar are
unsupported.

The are places where -2 is used for other meanings, or where it's
unclear if -2 signifies unsupported.  They are treated separately.