SERIALIZER: add functions for serialization to file
These functions are added:

- OSSL_SERIALIZER_CTX_set0_object()
- OSSL_SERIALIZER_CTX_get0_object()
- OSSL_SERIALIZER_CTX_set_do_output()
- OSSL_SERIALIZER_to_bio()
- OSSL_SERIALIZER_to_fp() (unless 'no-stdio')

OSSL_SERIALIZER_to_bio() and OSSL_SERIALIZER_to_fp() work as wrapper
functions, and simply just call the function specified with
OSSL_SERIALIZER_CTX_set_do_output() (let's call it the "do_output"
function) with the given serializer context and a BIO to output the
serialized result to.

The "do_output" function must have intimate knowledge of the object
given with OSSL_SERIALIZER_CTX_set0_object(), as that's the object to
be serialized.

This allows serializer initiators for any object to be defined
independently.